requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.25
pydantic~=1.10.14
PyJWT~=2.8.0
python-multipart==0.0.6
dnspython==2.5.0
email-validator==2.1.0.post1
idna==3.6
certifi==2024.2.2
cloudinary==1.38.0
six==1.16.0
urllib3==2.2.0
python-dotenv==1.0.1
